+++ test/support/schemas/keyed_event.ex
@@ -1,0 +1,10 @@
+defmodule QuackDB.TestSchemas.KeyedEvent do
+  @moduledoc false
+
+  use Ecto.Schema
+
+  schema "keyed_events" do
+    field(:name, :string)
+    field(:score, :integer)
+  end
+end